src/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pymongo==4.6.1
datetime==5.4